[#118] restructure schema statistics table model to separate logic from the Swing TableModel hierarchy
